I do the Graphic of the start of the app and implements the pokemons
